Revert "Authentication"
