+++ .gitignore
@@ -1,4 +1,4 @@
-node_modules/
+./node_modules/
 dist
-nyc_output/
+.nyc_output/
 coverage/
