+++ lib/models/menu.js
@@ -56,8 +56,6 @@ export default {
     return copy;
   },
   getMeals(offset = 0, limit = 10, search = '') {
-    offset = Number(offset);
-    limit = Number(limit);
     const result = this.dishes.filter(food => food.name.toLowerCase()
       .includes(search.toLowerCase()));
     return result.slice(offset, offset + limit);
