+++ lib/models/order.js
@@ -3,25 +3,30 @@ import menu from './menu';
 
 export default {
   orders: [],
-  addOrder(foodId, userId) {
+  addOrder(foodId) {
     const meal = menu.findMealById(foodId);
     if (!meal) { return undefined; }
-    const order = {
-      ...meal, userId, id: uniqId(), orderStatus: 'pending',
-    };
+    const order = { ...meal, id: uniqId(), orderStatus: 'pending' };
     this.orders.push(order);
     return order;
   },
   findOrderById(orderId) {
-    return this.orders.find(elem => elem.id === orderId);
+    const order = this.orders.find(elem => elem.id === orderId);
+    if (!order) { return undefined; }
+    return order;
   },
-  getOders(userId) {
-    return this.orders.filter(order => order.userId === userId);
+  getOders() {
+    return this.orders;
   },
   updateOrderStatus(orderId, orderStatus) {
-    return this.findOrderById(orderId).orderStatus = orderStatus;
+    const order = this.findOrderById(orderId);
+    if (!order) { return undefined; }
+    order.orderStatus = orderStatus;
+    return order;
   },
   deleteOrder(orderId) {
-    return this.orders.splice(this.orders.findIndex(order => order.id === orderId), 1);
+    const index = this.orders.findIndex(order => order.id === orderId);
+    if (index < 0) { return undefined; }
+    return this.orders.splice(index, 1);
   },
 };
