--- lib/models/user.js
@@ -1,23 +1,0 @@
-import bcrypt from 'bcryptjs';
-import uniqId from 'uniqid';
-
-export default {
-  records: [],
-  create(username, password, name) {
-    if (this.findByUsername(username)) { return false; }
-    const newUser = {};
-    newUser.username = username;
-    newUser.name = name;
-    newUser.id = uniqId();
-    bcrypt.genSalt(10, async (err, salt) => {
-      await bcrypt.hash(password, salt, (error, hash) => {
-        newUser.password = hash;
-        this.records.push(newUser);
-      });
-    });
-    return newUser;
-  },
-  findByUsername(username) {
-    return this.records.find(record => record.username === username);
-  },
-};
