+++ lib/routes/menu.js
@@ -7,6 +7,9 @@ const v1 = express.Router();
 
 // retrieve menu
 v1.get('/', middleware.getMenu, (req, res) => {
+  if (req.menu.length === 0) {
+    return res.status(404).send({ status: false, message: 'No dish found for the search criteria' });
+  }
   return res.status(200).send({ status: true, result: req.menu });
 });
