--- lib/routes/order.js
@@ -1,39 +1,0 @@
-import express from 'express';
-import middleware from '../middleware';
-
-const router = express();
-const v1 = express.Router();
-
-
-// Place order
-v1.post('/', middleware.verifyToken, middleware.createOrder, (req, res) => {
-  res.status(201).send({ status: true, result: req.order });
-});
-
-
-// Retrieve orders
-v1.get('/', middleware.verifyToken, middleware.getOrders, (req, res) => { 
-  res.status(200).send({ status: true, result: req.order });
-});
-
-// Retrieve specific order
-v1.get('/:id', middleware.verifyToken, middleware.getOrder, (req, res) => {
-  res.status(200).send({ status: true, result: req.order });
-});
-
-
-// Update order status
-v1.put('/:id', middleware.verifyToken, middleware.updateOrder, (req, res) => {
-  res.status(201).send({ status: true, result: req.order, message: `Order status has been updated to ${req.body.orderStatus}` });
-});
-
-
-// Delete order
-v1.delete('/:id', middleware.verifyToken, middleware.deleteOrder, (req, res) => {
-  res.status(200).send({ status: true, result: req.order, message: 'Order deleted' });
-});
-
-
-router.use('/api/v1/orders', v1);
-
-export default router;
