+++ package.json
@@ -7,7 +7,7 @@
     "start": "babel-node lib/app.js",
     "build": "babel lib -d dist",
     "serve": "node dist/app.js",
-    "test": "nyc --reporter=lcov mocha --require babel-core/register --require babel-polyfill --exit",
+    "test": "nyc --reporter=lcov mocha --require babel-core/register --exit",
     "coverage": "nyc report --reporter=text-lcov | coveralls"
   },
   "repository": {
@@ -21,12 +21,9 @@
   },
   "homepage": "https://github.com/Tonerolima/Fast-Food-Fast#readme",
   "dependencies": {
-    "bcryptjs": "^2.4.3",
     "body-parser": "^1.18.3",
     "express": "^4.16.3",
-    "jsonwebtoken": "^8.3.0",
     "method-override": "^3.0.0",
-    "multer": "^1.3.1",
     "uniqid": "^5.0.3"
   },
   "devDependencies": {
@@ -34,8 +31,6 @@
     "@babel/node": "^7.0.0",
     "ajv": "^6.5.3",
     "babel-cli": "^6.26.0",
-    "babel-core": "^6.26.3",
-    "babel-polyfill": "^6.26.0",
     "babel-preset-airbnb": "^2.6.0",
     "babel-preset-env": "^1.7.0",
     "babel-register": "^6.26.0",
@@ -47,6 +42,7 @@
     "eslint-plugin-import": "^2.14.0",
     "mocha": "^5.2.0",
     "mocha-lcov-reporter": "^1.3.0",
+    "nodemon": "^1.18.4",
     "nyc": "^13.0.1"
   }
 }
